+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 ### UNRELEASED
 
 * fix(webhooks): guarantee checksum uniqueness
+* fix(webhooks): meetingID not respected for scoped hooks
+* fix(webhooks): incorrectly sorted hooks in API responses
+* fix: external user ID missing from chat event
+* fix: internal/external-meeting-id missing from some events
 * test: add regression test for webhook checksum uniqueness
+* test: add regression tests for scoped hooks
 * build: address npm audit issues
 * build: bullmq@v4.18.3